Cleanup: Geometry Nodes: Use int instead of uint16_t
https://developer.blender.org/docs/handbook/guidelines/c_cpp/#integer-types

Pull Request: https://projects.blender.org/blender/blender/pulls/117594

+++ source/blender/blenkernel/intern/node_enum_definition.cc
@@ -82,7 +82,7 @@ void NodeEnumDefinition::clear()
   this->active_index = int16_t(active_index);
 }
 
-bool NodeEnumDefinition::move_item(const uint16_t from_index, const uint16_t to_index)
+bool NodeEnumDefinition::move_item(const int from_index, const int to_index)
 {
   if (to_index < this->items_num) {
     const int items_num = this->items_num;
